Fix Exercise_3 CFLAGS error

+++ embedded-linux/K26.3/truong-minh-khoa/session-02/Exercise_3/src/main.c
@@ -7,13 +7,13 @@
 
 int main() {
     char msg[MSG_SIZE];
-    int ret = 0;
+    size_t ret = 0;
     float a = 10.0f, b = 5.0f, c = 0.0f;
 
     ret = snprintf(msg, sizeof(msg), "calc_add: %.2f + %.2f = %.2f\n", a, b, calc_add(a, b));
     if(ret < 0 || ret >= sizeof(msg))
     {
-        fprintf(stderr, "Buffer overflow risk with ret:%d\n", ret);
+        fprintf(stderr, "Buffer overflow risk with ret:%lu\n", ret);
         return 1;
     }
     log_write(msg);
@@ -22,7 +22,7 @@ int main() {
     ret = snprintf(msg, sizeof(msg), "calc_sub: %.2f - %.2f = %.2f\n", a, b, calc_sub(a, b));
     if(ret < 0 || ret >= sizeof(msg))
     {
-        fprintf(stderr, "Buffer overflow risk with ret:%d\n", ret);
+        fprintf(stderr, "Buffer overflow risk with ret:%lu\n", ret);
         return 1;
     }
     log_write(msg);
@@ -31,7 +31,7 @@ int main() {
     ret = snprintf(msg, sizeof(msg), "calc_mul: %.2f * %.2f = %.2f\n", a, b, calc_mul(a, b));
     if(ret < 0 || ret >= sizeof(msg))
     {
-        fprintf(stderr, "Buffer overflow risk with ret:%d\n", ret);
+        fprintf(stderr, "Buffer overflow risk with ret:%lu\n", ret);
         return 1;
     }
     log_write(msg);
@@ -40,7 +40,7 @@ int main() {
     ret = snprintf(msg, sizeof(msg), "calc_div: %.2f / %.2f = %.2f\n", a, b, calc_div(a, b));
     if(ret < 0 || ret >= sizeof(msg))
     {
-        fprintf(stderr, "Buffer overflow risk with ret:%d\n", ret);
+        fprintf(stderr, "Buffer overflow risk with ret:%lu\n", ret);
         return 1;
     }
     log_write(msg);
@@ -51,7 +51,7 @@ int main() {
         ret = snprintf(msg, sizeof(msg), "Error - Division by zero: %.2f / %.2f = %.2f\n", a, c, calc_div(a, c));
         if(ret < 0 || ret >= sizeof(msg))
         {
-            fprintf(stderr, "Buffer overflow risk with ret:%d\n", ret);
+            fprintf(stderr, "Buffer overflow risk with ret:%lu\n", ret);
             return 1;
         }
         log_error(msg);
